Added new Transform (Chroma)
Also included Chord Progression functions. HMM_Stack_Transform, NMF_Stack_Transform

These take output from a transformation like CQT and a hyperparameter k (number of unique chords in progression) and output a condensed matrix representation. For example, if I have a progression of 5 chords (4 unique), the output will be an Y x 5 matrix (Y is the dimension of my transform. 60 if CQT) with each column being the different chord in sequence.

Maybe try this on some progressions and see how the model scores each column of the output matrix?